refactor(manifest/npm): break into mods
