+++ src/manifest/npm/commonjs.rs
@@ -1,0 +1,20 @@
+use manifest::npm::repository::Repository;
+
+#[derive(Serialize)]
+pub struct CommonJSPackage {
+    pub name: String,
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    pub collaborators: Vec<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+    pub version: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub license: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub repository: Option<Repository>,
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    pub files: Vec<String>,
+    pub main: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub types: Option<String>,
+}
