+++ src/manifest/npm/es6.rs
@@ -1,0 +1,20 @@
+use manifest::npm::repository::Repository;
+
+#[derive(Serialize)]
+pub struct ES6Package {
+    pub name: String,
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    pub collaborators: Vec<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+    pub version: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub license: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub repository: Option<Repository>,
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    pub files: Vec<String>,
+    pub module: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub types: Option<String>,
+}
