+++ src/manifest/npm/mod.rs
@@ -1,50 +1,12 @@
+mod commonjs;
+mod es6;
+pub mod repository;
+
+pub use self::commonjs::CommonJSPackage;
+pub use self::es6::ES6Package;
+
 #[derive(Serialize)]
 pub enum NpmPackage {
     CommonJSPackage(CommonJSPackage),
     ES6Package(ES6Package),
-}
-
-#[derive(Serialize)]
-pub struct CommonJSPackage {
-    pub name: String,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub collaborators: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    pub version: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub license: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub repository: Option<Repository>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub files: Vec<String>,
-    pub main: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub types: Option<String>,
-}
-
-#[derive(Serialize)]
-pub struct ES6Package {
-    pub name: String,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub collaborators: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    pub version: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub license: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub repository: Option<Repository>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub files: Vec<String>,
-    pub module: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub types: Option<String>,
-}
-
-#[derive(Serialize)]
-pub struct Repository {
-    #[serde(rename = "type")]
-    pub ty: String,
-    pub url: String,
 }
