+++ src/manifest/npm/repository.rs
@@ -1,0 +1,6 @@
+#[derive(Serialize)]
+pub struct Repository {
+    #[serde(rename = "type")]
+    pub ty: String,
+    pub url: String,
+}
